refactor exceptions to core/exceptions.py
