+++ ecom-backend/api/exception_handlers.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, Request
 from fastapi.responses import JSONResponse
-from exceptions import BaseAppException
+from core.exceptions import BaseAppException
 
 def app_exception_handler(request: Request, exc: BaseAppException):
     return JSONResponse(
