+++ ecom-backend/core/exceptions.py
