+++ ecom-backend/domains/users/api/dependencies.py
@@ -2,7 +2,7 @@ from api.dependencies import get_db
 from fastapi import Depends, Header
 from sqlalchemy.orm import Session
 from domains.users.models import UserDB
-from exceptions import UnauthorizedException, ResourceNotFoundException
+from core.exceptions import UnauthorizedException, ResourceNotFoundException
 
 
 # TODO: Make it depend on UserService
